requirements.txt
pandas
openpyxl
google-genai
streamlit
streamlit-tags
openai